requirements.txt
pdf2image=1.16.3
pytesseract=0.3.10
python-dotenv=1.0.0
openai=0.27.7
tiktoken=0.3.3
langchain=0.0.186